+++ lib/src/config/smart_config_attach.dart
@@ -26,6 +26,7 @@ class SmartConfigAttach {
       SmartNonAnimationType.highlightMask_nonAnimation,
     ],
     this.attachAlignmentType = SmartAttachAlignmentType.center,
+    this.backType = SmartBackType.normal,
     this.isExist = false,
   });
 
@@ -112,6 +113,7 @@ class SmartConfigAttach {
   ///
   /// true（返回事件将关闭loading，但是不会关闭页面），false（返回事件不会关闭loading，也不会关闭页面），
   /// 你仍然可以使用dismiss方法来关闭dialog
+  @Deprecated("please use backType")
   final bool backDismiss;
 
   /// Bind the dialog to the current page, the bound page is not on the top of the stack,
@@ -143,6 +145,12 @@ class SmartConfigAttach {
   ///
   /// attach dialog选择不同alignment属性时的对齐效果, 具体说明请参照[SmartAttachAlignmentType]
   final SmartAttachAlignmentType attachAlignmentType;
+
+  /// For different processing types of return events,
+  /// please refer to the description of [SmartBackType] for details
+  ///
+  /// 对于返回事件不同的处理类型, 具体可参照[SmartBackType]说明
+  final SmartBackType backType;
 
   /// whether attach dialog(showAttach()) exist on the screen
   ///