+++ lib/src/data/dialog_info.dart
@@ -1,14 +1,15 @@
 import 'dart:async';
 
 import 'package:flutter/material.dart';
+import 'package:flutter_smart_dialog/src/kit/typedef.dart';
 
+import '../../flutter_smart_dialog.dart';
 import '../helper/dialog_proxy.dart';
 import 'base_dialog.dart';
 
 class DialogInfo {
   DialogInfo({
     required this.dialog,
-    required this.backDismiss,
     required this.type,
     required this.tag,
     required this.permanent,
@@ -16,11 +17,11 @@ class DialogInfo {
     required this.bindPage,
     required this.route,
     required this.bindWidget,
+    required this.backType,
+    required this.onBack,
   });
 
   final BaseDialog dialog;
-
-  final bool backDismiss;
 
   final DialogType type;
 
@@ -37,4 +38,8 @@ class DialogInfo {
   final BuildContext? bindWidget;
 
   Timer? displayTimer;
+
+  final SmartBackType? backType;
+
+  final SmartOnBack? onBack;
 }