+++ lib/src/data/notify_info.dart
@@ -1,6 +1,7 @@
 import 'dart:async';
 
 import '../../flutter_smart_dialog.dart';
+import '../kit/typedef.dart';
 import 'base_dialog.dart';
 
 class NotifyInfo {
@@ -8,6 +9,7 @@ class NotifyInfo {
     required this.dialog,
     required this.tag,
     required this.backType,
+    required this.onBack,
   });
 
   final BaseDialog dialog;
@@ -17,4 +19,6 @@ class NotifyInfo {
   SmartBackType backType;
 
   Timer? displayTimer;
+
+  final SmartOnBack? onBack;
 }