+++ lib/src/kit/typedef.dart
@@ -1,1 +1,5 @@
+import 'dart:async';
+
 typedef FutureVoidCallback = Future<void> Function();
+
+typedef SmartOnBack = FutureOr<bool> Function();